List functions added to sml.

+++ Monty/src/main/java/sml/casts/BooleanToFloat.java
@@ -20,8 +20,8 @@ public class BooleanToFloat {
 
 	public static Float booleanToFloat(Boolean bool) {
 		if (bool == true)
-			return (Float) 1f;
-		return (Float) 0f;
+			return 1f;
+		return 0f;
 	}
 
 }
