+++ Monty/src/main/java/sml/casts/StringToFloat.java
@@ -24,7 +24,7 @@ public class StringToFloat {
 		if (str.matches("[+-]?[0-9]+\\.[0-9]+"))
 			return Float.parseFloat(str);
 		else if (str.matches("[+-]?[0-9]+"))
-			return (Float) (float) Integer.parseInt(str);
+			return (float) Integer.parseInt(str);
 		else
 			new MontyException("Unknown number format for float type:\t" + str);
 		return null;
